+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Wrong parameter setting procedure in TcpServer::SetKeepAliveIdleTime.
+- WiFi station and Ethernet initialization rollback.
 
 ## [1.1.2] - 2024-09-26
 ### Added